Title: 소수점 반올림하기, Time: 116ms, Memory: 8MB, Status: Passed - CodeTree

+++ 231214/소수점 반올림하기/rounding-decimal-points.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
-        double a = 25.352;
+        float a = 25.352f;
 
         System.out.printf("%.1f", a);
     }
